+++ NEWS.md
@@ -4,6 +4,7 @@
 
   * New options `heemod.verbose` and `heemod.memotime`.
   * More informative messages, especially in verbose mode.
+  * Added option to pool female and male mortality rates in WHO data.
 
 # heemod 0.4.0